Support copying data from p2p buffers.
Sometimes, we may want not want load the program or the data using DMA.
Instead, we might want to use P2PDMA to move the data into the device.

In this case, however, the program/data will be somewhere in the p2p
buffer and we need to move it to the appopriate locations in the device.

This patch introduces new operations to do just that.

Note that this patch DOES NOT do the P2PDMA part. This must be done
somewhere else.

Signed-off-by: Martin Ichilevici de Oliveira <martin.i.oliveira@gmail.com>